requirements.txt
Django==4.1.7
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2-binary==2.9.5
djangorestframework-simplejwt==5.2.2
